requirements.txt
# FastAPI
fastapi>=0.115.0
uvicorn>=0.30.0
pydantic>=2.0.0

# LangChain & AI
langchain-core>=0.3.0
langchain-community>=0.3.0
langchain-text-splitters>=0.3.0
langchain-google-genai>=2.0.0
langchain-redis>=0.2.0
langgraph>=0.2.0

# Vector Store & Cache
pinecone>=5.0.0
redis>=5.0.0

# Document Loading
beautifulsoup4>=4.12.0
lxml>=5.0.0

# Environment
python-dotenv>=1.0.0

# Observability (optional)
langfuse>=2.0.0
